+++ lib/rfacter/config/settings.rb
@@ -23,16 +23,16 @@ class RFacter::Config::Settings
   #   schemes to use when contacting them.
   attr_reader :nodes
 
-  # A boolean switch for measuring fact resolution time
+  # A boolean switch for enabling execution tracing
   #
   # @return [Boolean] Defaults to false.
-  attr_accessor :timing
+  attr_accessor :trace
 
   def initialize(**options)
     @logger = RFacter::Util::Logger.new($stderr)
     @logger.level = Logger::WARN
 
-    @timing = false
+    @trace = false
     @nodes = Hash.new
   end
 end